requirements.txt
Django==3.1.2
django-debug-toolbar==3.1.1
django-filter==2.4.0
django-import-export==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.1

requests==2.24.0
